+++ app/src/main/java/com/jingom/seizetheday/domain/usecase/GetThanksRecordsPagingDataFlowUseCase.kt
@@ -9,7 +9,7 @@ import javax.inject.Inject
 class GetThanksRecordsPagingDataFlowUseCase @Inject constructor(
 	private val thanksRecordRepository: ThanksRecordRepository
 ) {
-	operator fun invoke(startThanksId: Int? = null): Flow<PagingData<ThanksRecord>> {
+	operator fun invoke(startThanksId: Long? = null): Flow<PagingData<ThanksRecord>> {
 		return thanksRecordRepository.getThanksRecordsPagingFlow(startThanksId)
 	}
 }
